requirements.txt
flask>=2.0.0
nltk>=3.6.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
numpy>=1.19.0
pandas>=1.1.0
pymongo>=4.0.0
dnspython>=2.0.0
python-dotenv>=0.19.0
bcrypt>=4.0.0
flask-socketio>=5.3.0
redis>=4.5.0
email-validator>=2.0.0
gunicorn>=20.1.0
eventlet>=0.33.0
